Add support for Babel legacy decorators
This commit also runs the tracked decorator tests against both Babel and
TypeScript output.
