+++ package.json
@@ -21,7 +21,6 @@
     "@babel/core": "^7.2.2",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-proposal-decorators": "^7.4.4",
-    "@babel/plugin-transform-modules-amd": "^7.5.0",
     "@babel/plugin-transform-proto-to-assign": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
     "@types/qunit": "^2.0.31",
