+++ packages/@glimmer/tracking/test/browser/helpers/tags.ts
@@ -1,0 +1,11 @@
+import { bump, Tag } from '@glimmer/reference';
+
+export function assertValidAfterUnrelatedBump(tag: Tag, snapshot: number) {
+  bump();
+
+  QUnit.assert.strictEqual(
+    tag.validate(snapshot),
+    true,
+    'tag is still valid after an unrelated bump'
+  );
+}
